`fn get_random_number`: Make `r` var a `c_uint`.

+++ src/filmgrain.rs
@@ -201,9 +201,9 @@ macro_rules! decl_fguv_32x32xn_fn {
 
 #[inline]
 fn get_random_number(bits: c_int, state: &mut c_uint) -> c_int {
-    let r = *state as c_int;
-    let bit = ((r >> 0 ^ r >> 1 ^ r >> 3 ^ r >> 12) & 1) as c_uint;
-    *state = (r >> 1) as c_uint | bit << 15;
+    let r = *state;
+    let bit = (r >> 0 ^ r >> 1 ^ r >> 3 ^ r >> 12) & 1;
+    *state = (r >> 1) | bit << 15;
 
     (*state >> 16 - bits & (1 << bits) - 1) as c_int
 }
